requirements.txt
anytree==2.8.0
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
Beaker==1.11.0
btrfsutil==5.18.1
CacheControl==0.12.6
catfish==4.16.3
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
colorama==0.4.5
commonmark==0.9.1
contextlib2==21.6.0
cryptography==37.0.2
distlib==0.3.4
distro==1.7.0
Django==4.0.6
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.19
filelock==3.7.1
future==0.18.2
gufw==22.4.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
keyutils==0.6
lensfun==0.3.3
LibAppArmor==3.0.4
libtorrent==2.0.6
lightdm-gtk-greeter-settings==1.2.2
lit==14.0.6.dev0
louis==3.22.0
lxml==4.9.0
Mako==1.1.6
manjaro-sdk==0.1
Markdown==3.3.7
MarkupSafe==2.1.1
menulibre==2.2.3
meson==0.63.0
more-itertools==8.13.0
msgpack==1.0.3
mugshot==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pep517==0.12.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
progress==1.6
psutil==5.9.1
ptyprocess==0.7.0
pycairo==1.21.0
pycparser==2.21
Pygments==2.12.0
PyGObject==3.42.1
pyinotify==0.9.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
python-distutils-extra==2.39
python-Levenshtein==0.12.2
pytz==2022.1
pyxdg==0.28
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.10
requests==2.27.1
resolvelib==0.8.1
retrying==1.3.3
rgw==2.0.0
rich @ file:///build/python-rich/src/rich-12.4.4/dist/rich-12.4.4-py3-none-any.whl
shiboken6==6.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
team==1.0
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.6.26
typing_extensions==4.2.0
udiskie==2.4.2
ufw==0.36.1
ulauncher==5.14.5
urllib3==1.26.9
validate-pyproject==0.9
virtualenv==20.15.1
webencodings==0.5.1
websocket-client==1.3.3
zipp==3.8.0
